fix rest serialization, `null` in rest upsert and delete (#43)
* Fix a a JSON encode error. `.map` returns a `MappedListIterable` which `jsonEncode` cannot parse. It can parse `List<dynamic>`.
* When `url` is `null` in `RestProvider#upsert` and `RestProvider#delete`, return `null` and do not attempt to perform the action

+++ packages/brick_rest/lib/rest.dart
@@ -43,7 +43,9 @@ class RestProvider implements Provider<RestModel> {
   @override
   Future<http.Response> delete<_Model extends RestModel>(instance, {query, repository}) async {
     final url = urlForModel<_Model>(query, instance);
+    if (url == null) return null;
     _logger.fine('DELETE $url');
+
     final resp = await client.delete(url, headers: headersForQuery(query));
 
     _logger.finest('caller=delete url=$url statusCode=${resp?.statusCode} body=${resp?.body}');
@@ -105,6 +107,8 @@ class RestProvider implements Provider<RestModel> {
     final body = await adapter.toRest(instance, provider: this, repository: repository);
 
     final url = urlForModel<_Model>(query, instance);
+    if (url == null) return null;
+
     final resp = await _sendUpsertResponse(url, body, query, adapter.toKey);
 
     _logger.finest('caller=upsert url=$url statusCode=${resp?.statusCode} body=${resp?.body}');
